feat: 인벤토리 버리기 기능 추가
버튼 선택 시 하이라이트
재 클릭 시 선택 취소
인벤토리 창 껐다 키면 버튼 초기화
제거 버튼 추가
용량 제한 추가
